refactor: use mod.rs across the codebase
- Previously all module structures were defined in main.rs  This had become unweildly

- Now using mods.rs
