+++ src/callbacks/mod.rs
@@ -1,0 +1,4 @@
+pub mod delete_container;
+pub mod delete_image;
+
+pub use delete_container::DeleteContainer;
