+++ src/components/mod.rs
@@ -1,0 +1,7 @@
+pub mod alert_modal;
+pub mod boolean_modal;
+pub mod footer;
+pub mod header;
+pub mod help;
+pub mod input_field;
+pub mod resize_notice;
