+++ src/pages/mod.rs
@@ -1,0 +1,4 @@
+pub mod attach;
+pub mod containers;
+pub mod images;
+pub mod logs;
