+++ src/traits/mod.rs
@@ -1,0 +1,8 @@
+mod callback;
+mod component;
+mod page;
+
+pub use callback::Callback;
+pub use component::Component;
+pub use component::ModalComponent;
+pub use page::Page;
