+++ src/ui/mod.rs
@@ -1,0 +1,4 @@
+pub mod app;
+pub mod page_manager;
+
+pub use app::App;
